Create separate lessons view and views.go
Reviewer: none
Issue: ME-25
